requirements.txt
# Core Django requirements
Django==5.1.2
djangorestframework==3.14.0
django-cors-headers==3.13.0

# Database driver
psycopg2-binary==2.9.9

# Other dependencies
pillow==9.4.0
django-filter==22.1
django-crispy-forms==1.14.0
crispy-bootstrap4==2024.1
django-allauth==0.52.0
google ==3.0.0
django-widget-tweaks==1.4.12
celery==5.2.7
redis==4.4.0
stripe==5.0.0
django-anymail[mailgun]==8.6
django-htmx ==1.21.0
# AI and Machine Learning
openai==1.3.0
transformers==4.35.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
# Video Streaming and Real-time Communication
channels==4.0.0
channels-redis==4.1.0
django-storages==1.14.2
opencv-python-headless==4.8.1.78

# Search and Recommendation
elasticsearch-dsl==7.4.1
django-elasticsearch-dsl==7.3.0

# Payment Integration (for African payment gateways)
rave-python==1.2.16  # Flutterwave Python SDK

# Performance and Caching
django-cacheops==7.0.1
django-redis==5.4.0

# API and Documentation
drf-yasg==1.21.7
django-rest-swagger==2.2.0

# Testing and Development
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0